# [Basic] Make OutputByteStream subclassable