--- a/Sources/Build/describe().swift
+++ b/Sources/Build/describe().swift
@@ -102,7 +102,7 @@ public func describe(_ prefix: AbsolutePath, _ conf: Configuration, _ graph: Pac
 }
 
 private func write(path: AbsolutePath, write: (OutputByteStream) -> Void) throws -> AbsolutePath {
-    let stream = OutputByteStream()
+    let stream = BufferedOutputByteStream()
     write(stream)
     try localFileSystem.writeFileContents(path, bytes: stream.bytes)
     return path