--- a/Sources/Commands/init.swift
+++ b/Sources/Commands/init.swift
@@ -15,7 +15,7 @@ import POSIX
 private extension FileSystem {
     /// Write to a file from a stream producer.
     mutating func writeFileContents(_ path: AbsolutePath, body: @noescape (OutputByteStream) -> ()) throws {
-        let contents = OutputByteStream()
+        let contents = BufferedOutputByteStream()
         body(contents)
         try createDirectory(path.parentDirectory, recursive: true)
         try writeFileContents(path, bytes: contents.bytes)