--- a/Tests/Basic/ByteStringTests.swift
+++ b/Tests/Basic/ByteStringTests.swift
@@ -63,7 +63,7 @@ class ByteStringTests: XCTestCase {
     }
 
     func testByteStreamable() {
-        let s = OutputByteStream()
+        let s = BufferedOutputByteStream()
         s <<< ByteString([1, 2, 3])
         XCTAssertEqual(s.bytes, [1, 2, 3])
     }