This makes the OutputByteStream class subclassable and implements an
InMemory version.