@aciidb0mb3r 2016-07-20
Performance data: (Keep in mind this copies data into a new buffer for now)

| Test | Old (sec) | This PR (sec) |
| --- | :---: | :---: |
|test1MBOf16ByteArrays_X100 			|	0.291	|0.351|
|test1MBOf16ByteArraySlice_X100 		|	0.307	|0.360|
|test1MBOf16ByteStrings_X10 			|	0.113	|0.108|
|test1MBOf1KByteArrays_X1000			|	0.184	|0.245|
|test1MBOf1KByteStrings_X100			|	0.558	|0.569|
|test1MBOfByte_X10  					|	0.245	|0.245|
|test1MBOfCharacters_X1					|	0.401	|0.363|
|test1MBOfJSONEncoded16ByteStrings_X10	|	0.112	|0.139|
|testFormattedJSONOutput				|	0.177	|0.214|