@abertelrud 2016-07-20
I think that `BufferedOutputByteStream` and `OutputByteStream` are the best despite being long, given that `OutputStream` is defined to be a text stream (it would have been nice to have `OutputStream` be in terms of raw data and then an `OutputTextStream` on top of that, but ...)