@ddunbar 2016-07-21
Swift's `OutputStream` just got renamed to `TextOutputStream` so `OutputStream` is technically more open to us...