@abertelrud 2016-07-21
That is fantastic news, from a conceptual-clarity point-of-view.  In that case, I think we should have OutputStream and BufferedOutputStream.