@ddunbar 2016-07-21
Okay, those names work for me. Let's land this change, then rename.