@aciidb0mb3r 2016-07-21
Updated
New stats: 

| Test | current code (sec) | This PR (sec) |
| --- | :---: | :---: |						
|test1MBOf16ByteArrays_X100 			|   0.287	| 0.464
|test1MBOf16ByteArraySlice_X100 		|   0.340	| 0.491
|test1MBOf16ByteStrings_X10 			|   0.096	| 0.112
|test1MBOf1KByteArrays_X1000			|   0.198	| 0.253
|test1MBOf1KByteStrings_X100			|   0.576	| 0.583
|test1MBOfByte_X10  					|   0.217	| 0.267
|test1MBOfCharacters_X1					|   0.419	| 0.383
|test1MBOfJSONEncoded16ByteStrings_X10	|   0.103	| 0.133
|test1MBOfSequence_X10					|   0.148	| 0.172
|testFormattedJSONOutput				|   0.204	| 0.229