@ddunbar 2016-07-21
Looks pretty good, a couple minor comments inline. I think we should also extend the correctness test coverage to try and make sure we have coverage of the various ways that writes can straddle the internal buffer. So things like write a small chunk then a big chunk, write a single big chunk, write a big chunk then a small chunk.