@aciidb0mb3r 2016-07-21
Added a test