@ddunbar 2016-07-21
@aciidb0mb3r and I discussed this on Slack, we are going to land this current version then continue to iterate on the performance...